Rename component, update imports/export with a new name
