+++ packages/ui/src/components/va-tree-view/index.ts
@@ -1,7 +1,7 @@
 import withConfigTransport from '../../services/config-transport/withConfigTransport'
-import VaTreeRoot from './VaTreeRoot.vue'
+import VaTreeView from './VaTreeView.vue'
 
 export { default as VaTreeNode } from './VaTreeNode'
 export { default as VaTreeCategory } from './VaTreeCategory'
 
-export default withConfigTransport(VaTreeRoot as any)
+export default withConfigTransport(VaTreeView as any)
